New function kaa.delay(secs) that returns an InProgress object that finishes after the given number of seconds; Renamed InProgress.link() to InProgress.waitfor(); Signal.async() changed to Signal.__inprogress__, and a new function inprogress() created to invoke __inprogress__ on any object (similar to len() and __len__).
This new protocol is so that one can pass arbitrary objects within kaa where an
InProgress is supported, and it will be translated or coerced to an InProgress.
One example is Signal (which is the only object that has an __inprogress__
implemented.)  So you can now pass signals directly to InProgress(Any|All)
and can now yield signals in coroutines.

Signal.__inprogress__ returns an InProgress object that is weakly connected to
the signal, because the existence of the Signal isn't any reason to keep the
InProgress alive.  Weakly connecting means when the InProgress goes away, it
will be cleanly disconnected from the Signal.

Global references to active CoroutineInProgress objects are now kept;
see CoroutineInProgress.__init__ for detailed rationale.  Warnings are
now logged when InProgressSignals or InProgressList is used as they are
now considered deprecated.


git-svn-id: svn://svn.freevo.org/kaa/trunk@3424 a8f5125c-1e01-0410-8897-facf34644b8e