New custom sphinx directives:
autosynopsis: adds 3 summary tables for methods, properties, and signals
automethods: lists all methods in the class, preceded by a 'Methods' header
autosignals: same as above but for signals
autoproperties: same as above but for properties

The autosynopsis directive works, but it's a serious kludge.

See io.rst for a demo of the new directives.



git-svn-id: svn://svn.freevo.org/kaa/trunk@3770 a8f5125c-1e01-0410-8897-facf34644b8e