async.py: new exception InProgressAborted, which an InProgress is finished with if abort() is called.  If an InProgress A is waiting on InProgress B, and B is finished with InProgressAborted, then A aborts and emits its 'abort' signal so that abort handlers can clean up.
InProgress.timeout() now takes an abort kwarg which will abort the original
InProgress if there's a timeout.  (This is just a convenience.)

thread.py: implement abort() for ThreadInProgress.  This is a bit magical and
untested.

coroutine.py: if coroutine A yields coroutine B and A is stopped for some reason,
we don't abort B because someone else may also be waiting on B.  However if A
is explicitly aborted, then we do abort B.  (This logic is precarious and I need
to test it better.  I just wanted to commit what I had before I forgot.)



git-svn-id: svn://svn.freevo.org/kaa/trunk@3962 a8f5125c-1e01-0410-8897-facf34644b8e