Factor out code actually common to Dict and List into base class Container, and subclass Dict and List from that; dict and list now behave a lot more like their respective datatypes.  It's now possible to make assignments:
   config.dict = {'foo': 42, 'bar': 84}
   config.list = ['apple', 'banana', 'carrot']

Indexes in lists are now stripped, so the above list is outputted as:

   list[] = apple
   list[] = banana
   list[] = carrot

If an index is present in the config, it will still parse correctly IFF
the index is one more than the current length.  For example, the first 3
lines are ok, but the 4th line will fail:

    list[0] = apple
    list[1] = banana
    list[2] = carrot
    list[42] = date

This is unlikely to break any existing config files.  Now the prefered format
when hand-editing is to leave the index off, which is easier for humans anyway.



git-svn-id: svn://svn.freevo.org/kaa/trunk@4217 a8f5125c-1e01-0410-8897-facf34644b8e